[Core] Work around libtorrent size limit
When loading a torrent_info by file path, libtorrent has a hardcoded
size limit that some torrents exceed. This means that the torrent can be
added normally to deluge, but on restart it fails to load the file from
the state directory and the torrent is lost.

Work around the hardcoded buffer size limit by reading in the file
directly and passing the buffer.

See: https://libtorrent-discuss.narkive.com/wYACQDb7/libtorrent-loading-a-big-torrent-fails
Fixes: https://forum.deluge-torrent.org/viewtopic.php?t=56856
Closes: https://github.com/deluge-torrent/deluge/pull/483