fix(gtkui): resolve password and auth dialogs not handling user input
A previous fix (1e3c6246) changed the order of destroying the dialog to
preceed the response callback. This was to avoid multiple dialogs being
shown if waiting to destroy dialog after callback.

The problem was that several auth dialogs that use BaseDialog relied on
accessing the dialog entry properties in the callback but with dialog
destroyed were no longer available.

A quick simple solution is to return these properties in the callback.